Varios y FIX #199 Limitar comandos - no todos. - Limitar comandos con un mínimo de mensajes siguientes escritos para volver a ejecutar el comando. Menos spam. No todos lo harán igualmente. - TODO del plugin voz a texto, cancelar si no existe el programa ffmpeg y demás dependencias. - Ayuda del comando /iv - Algo de limpieza.

+++ application/plugins/voice_text.php
@@ -1,7 +1,9 @@
 <?php
 
 if($telegram->voice() && $telegram->voice(TRUE)->duration <= 7){
-    // $telegram->send->text("Recibo voz de " .$telegram->voice(TRUE)->duration ." segundos.")->send();
+	if(!file_exists("ffmpeg")){ return; } // TODO FIXME
+	// Falta indicar dependencia de las apps.
+
     $tmp = tempnam("/tmp", "vrec");
 
     file_put_contents($tmp, file_get_contents($telegram->download($telegram->voice())));
